refactor(callbacks): extract Weather tab fast path to _callbacks_weather.py
Step 6 of the callbacks.py decomposition (issue #87). Continues the
per-tab split established by PRs #98 / #99 / #100 / #101 / #102.

## What moves

``_weather_tab_from_redis`` (~134 lines) — the Redis fast path that
builds the six Plotly figures (temp/wind/solar scatters, correlation
heatmap, feature-importance bar chart, seasonal decomposition subplots)
for the Weather-Energy Correlation tab from the scoring job's hourly
``wattcast:weather-correlation:{region}`` payload — moves out of
callbacks.py into ``components/_callbacks_weather.py``.

## Orphaned-callback observation

Same situation as ``_generation_tab_from_redis`` (#102): tested but
no longer wired into ``register_callbacks``. Kept the re-export so
tests stay valid; flagged with explicit ``# noqa`` so the next pass
can decide delete vs rewire. The Generation and Weather paths likely
share a common root cause from an earlier refactor that removed the
calling callback decorator without removing the helper bodies.

## Public-import surface preserved

``components/callbacks.py`` re-imports ``_weather_tab_from_redis`` by
name with an explicit ``# noqa: F401 — re-export`` marker.

## Test patches updated

5 ``@patch("components.callbacks.redis_get")`` decorators inside
``TestWeatherTabFromRedis`` rewired to ``components._callbacks_weather``.

## Size impact

callbacks.py: 5,802 → 5,668 (–134 lines)
_callbacks_weather.py: new, 191 lines

Cumulative progress on the decomposition:
* PR #98 (shared):      7,167 → 7,051 (–116)
* PR #99 (us_grid):     7,051 → 6,389 (–662)
* PR #100 (models):     6,389 → 6,183 (–206)
* PR #101 (alerts):     6,183 → 5,978 (–205)
* PR #102 (generation): 5,978 → 5,802 (–176)
* this PR (weather):    5,802 → 5,668 (–134)

Total: −1,499 lines from the monolith across 6 PRs (−20.9%).

## Tests

All 1,568 unit tests pass locally. Ruff clean.
